fixed typo

+++ utilities/ContractorNuc.cpp
@@ -465,7 +465,7 @@ int main(int argc, char* argv[])
                         thread_for(j,ref.dimension(2),{
                             for (unsigned int i = 0; i < ref.dimension(1); ++i)
                             {
-                                lastTerm[t](mu, i, j, k) = ref(mu, i, j, k);
+                                lastTerm[t]((unsigned long long)mu, i, j, k) = ref((unsigned long long)mu, i, j, k);
                             }
                         });
                     });
